pyproject.toml
[project]
name = "bimap"
version = "0.1.0"
authors = [{name="Moritz Palm", email="moritz.palm@fau.de"}]
description = "Nonrigid Image Segmentation for Calcium Imaging Data"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.11"
dependencies = [
    "antspyx>=0.5.3",
    "hatchling>=1.27.0",
    "imageio[ffmpeg]>=2.37.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.3",
    "lab>=8.4",
    "matplotlib>=3.10.3",
    "nibabel>=5.3.2",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "polars>=1.29.0",
    "ruff>=0.11.12",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "voxelmorph>=0.2",
]

[project.urls]
Repository = "https://github.com/MoritzPalm/BIMAP"
Issues = "https://github.com/MoritzPalm/BIMAP/issues"

[build-system]
requires = ["hatchling >= 1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["src/**"]

[tool.ruff]
line-length = 120
target-version = "py313"
#excluding imported code
extend-exclude = ["torch_lddmm.py"]
exclude = ["voxelmorph"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["T201"]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[tool.hatch.metadata]
allow-direct-references = true



